requirements.txt
# Voice Agent Dependencies

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# WebSocket support
websockets>=12.0

# HTTP client
httpx>=0.26.0
aiohttp>=3.9.0

# Telnyx SDK
telnyx>=2.1.0

# Deepgram SDK
deepgram-sdk>=3.0.0

# ElevenLabs (HTTP, no official SDK needed)
# Using httpx for API calls

# Audio processing
numpy>=1.26.0
# Note: audioop is built-in for Python <3.13

# Async utilities
asyncio-throttle>=1.0.0

# Configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Security
cryptography>=42.0.0  # For webhook signature verification

# Logging
structlog>=24.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
twilio>=9.0.0

# Voice Agent V2.0 - Streaming & Interruption
anthropic>=0.18.0  # Claude streaming API
tenacity>=8.2.0  # Retry logic with exponential backoff
pybreaker>=1.0.0  # Circuit breaker pattern
prometheus-client>=0.19.0  # Metrics collection
redis>=5.0.0  # Optional: state persistence
